+++ R/zzz.R
@@ -3,6 +3,10 @@
 
 .onAttach <- function(libname, pkgname){
   packageStartupMessage(
-    "Be sure to set the download folder using `prism_set_dl_dir()`."
+    "Be sure to set the download folder using `prism_set_dl_dir()`.",
+    "\nBy default, prism data are downloaded in GeoTIFF format. ",
+    "See ?prism_set_format for more details."
   )
+  
+  options(prism.format = "geotiff")
 }