+++ NEWS.md
@@ -6,6 +6,8 @@
 - deleted `keep_pre81_months` paramter from `get_prism_annual()` and `get_prism_monthlys()`. Deprecated in v0.3.0
 - deleted `get_prism_station_md()`, `prism_slice()`, `prism_stack()`, `del_early_prov()`, `ls_prism_data()`, `check_corrupt()`, `path_check()`, `prism_check()`. Deprecated in version XXX.
 - removed dependency on magrittr package; now depends on R >= 4.1.0. replaced %>% with |>.
+- adding new download formats (#131)
+  - added `prism_set_format()` and `prism_get_format()`
 
 # prism 0.3.0